tweak test_type_info:  turns out looking for "long" is not consistent across compilers/platforms (some use "long int")

+++ newtestsuite/src/symtab/test_type_info_mutatee.c
@@ -48,7 +48,7 @@ enum enum1 {
 
 struct mystruct {
 	int elem1;
-	long elem2;
+	double elem2;
 	char elem3;
 	float elem4;
 } ms;
